pyproject.toml
[project]
name = "comms"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14,<4"
dependencies = [
    "django",
    "django-environ",
    "psycopg",
    "whitenoise",
    "sentry-sdk[django]",
    "gunicorn",
    "djangorestframework",
    "django-rest-knox",
    "httpx",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8",
    "flake8-pyproject",
    "flake8-isort",
    "pytest",
    "pytest-django"
]

[tool.flake8]
exclude = [
    ".heroku",
    ".local",
]
ignore = [
    "E501",  # line too long
]

[tool.isort]
skip_gitignore = true
lines_after_imports = 2
# 5 = Hanging Grid Grouped
multi_line_output = 5
include_trailing_comma = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "comms.settings"
FAIL_INVALID_TEMPLATE_VARS = true
python_files = ["tests.py", "test_*.py", "*_tests.py"]
